data type rethink

+++ app/src/core/components/gallery/gallery_settings/thumbnail_size_slider.rs
@@ -2,13 +2,13 @@ pub mod message;
 pub mod update;
 pub mod view;
 
-const DEFAULT_THUMBNAIL_SIZE: u32 = 224;
-const MIN_THUMBNAIL_SIZE: u32 = 64;
-const MAX_THUMBNAIL_SIZE: u32 = 448;
-const SLIDER_STEP: u32 = 32;
+const DEFAULT_THUMBNAIL_SIZE: u16 = 224;
+const MIN_THUMBNAIL_SIZE: u16 = 64;
+const MAX_THUMBNAIL_SIZE: u16 = 448;
+const SLIDER_STEP: u16 = 32;
 
 pub struct ThumbnailSizeSlider {
-    pub value: u32,
+    pub value: u16,
 }
 
 impl Default for ThumbnailSizeSlider {
